+++ pyVHDLModel/Exception.py
@@ -1,6 +1,43 @@
+# ==================================================================================================================== #
+#             __     ___   _ ____  _     __  __           _      _                                                     #
+#   _ __  _   \ \   / / | | |  _ \| |   |  \/  | ___   __| | ___| |                                                    #
+#  | '_ \| | | \ \ / /| |_| | | | | |   | |\/| |/ _ \ / _` |/ _ \ |                                                    #
+#  | |_) | |_| |\ V / |  _  | |_| | |___| |  | | (_) | (_| |  __/ |                                                    #
+#  | .__/ \__, | \_/  |_| |_|____/|_____|_|  |_|\___/ \__,_|\___|_|                                                    #
+#  |_|    |___/                                                                                                        #
+# ==================================================================================================================== #
+# Authors:                                                                                                             #
+#   Patrick Lehmann                                                                                                    #
+#                                                                                                                      #
+# License:                                                                                                             #
+# ==================================================================================================================== #
+# Copyright 2017-2023 Patrick Lehmann - Boetzingen, Germany                                                            #
+# Copyright 2016-2017 Patrick Lehmann - Dresden, Germany                                                               #
+#                                                                                                                      #
+# Licensed under the Apache License, Version 2.0 (the "License");                                                      #
+# you may not use this file except in compliance with the License.                                                     #
+# You may obtain a copy of the License at                                                                              #
+#                                                                                                                      #
+#   http://www.apache.org/licenses/LICENSE-2.0                                                                         #
+#                                                                                                                      #
+# Unless required by applicable law or agreed to in writing, software                                                  #
+# distributed under the License is distributed on an "AS IS" BASIS,                                                    #
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.                                             #
+# See the License for the specific language governing permissions and                                                  #
+# limitations under the License.                                                                                       #
+#                                                                                                                      #
+# SPDX-License-Identifier: Apache-2.0                                                                                  #
+# ==================================================================================================================== #
+#
+"""
+This module contains parts of an abstract document language model for VHDL.
+
+Exceptions contains all structured errors.
+"""
 from pyTooling.Decorators import export
 
-from pyVHDLModel import VHDLModelException, Symbol
+from pyVHDLModel import VHDLModelException
+from pyVHDLModel.Symbol import NewSymbol
 
 
 @export
@@ -276,10 +313,10 @@ class ReferencedLibraryNotExistingError(VHDLModelException):
 	Message: :pycode:`f"Library '{librarySymbol.Identifier}' referenced by library clause of context '{context.Identifier}' doesn't exist in design."`
 	"""
 
-	_librarySymbol: Symbol
+	_librarySymbol: NewSymbol
 	_context: 'Context'
 
-	def __init__(self, context: 'Context', librarySymbol: Symbol):
+	def __init__(self, context: 'Context', librarySymbol: NewSymbol):
 		"""
 		Initializes the exception message based on given context and library objects.
 
@@ -291,7 +328,7 @@ class ReferencedLibraryNotExistingError(VHDLModelException):
 		self._context = context
 
 	@property
-	def LibrarySymbol(self) -> Symbol:
+	def LibrarySymbol(self) -> NewSymbol:
 		return self._librarySymbol
 
 	@property
