+++ pyVHDLModel/Expression.py
@@ -1,8 +1,45 @@
+# ==================================================================================================================== #
+#             __     ___   _ ____  _     __  __           _      _                                                     #
+#   _ __  _   \ \   / / | | |  _ \| |   |  \/  | ___   __| | ___| |                                                    #
+#  | '_ \| | | \ \ / /| |_| | | | | |   | |\/| |/ _ \ / _` |/ _ \ |                                                    #
+#  | |_) | |_| |\ V / |  _  | |_| | |___| |  | | (_) | (_| |  __/ |                                                    #
+#  | .__/ \__, | \_/  |_| |_|____/|_____|_|  |_|\___/ \__,_|\___|_|                                                    #
+#  |_|    |___/                                                                                                        #
+# ==================================================================================================================== #
+# Authors:                                                                                                             #
+#   Patrick Lehmann                                                                                                    #
+#                                                                                                                      #
+# License:                                                                                                             #
+# ==================================================================================================================== #
+# Copyright 2017-2023 Patrick Lehmann - Boetzingen, Germany                                                            #
+# Copyright 2016-2017 Patrick Lehmann - Dresden, Germany                                                               #
+#                                                                                                                      #
+# Licensed under the Apache License, Version 2.0 (the "License");                                                      #
+# you may not use this file except in compliance with the License.                                                     #
+# You may obtain a copy of the License at                                                                              #
+#                                                                                                                      #
+#   http://www.apache.org/licenses/LICENSE-2.0                                                                         #
+#                                                                                                                      #
+# Unless required by applicable law or agreed to in writing, software                                                  #
+# distributed under the License is distributed on an "AS IS" BASIS,                                                    #
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.                                             #
+# See the License for the specific language governing permissions and                                                  #
+# limitations under the License.                                                                                       #
+#                                                                                                                      #
+# SPDX-License-Identifier: Apache-2.0                                                                                  #
+# ==================================================================================================================== #
+#
+"""
+This module contains parts of an abstract document language model for VHDL.
+
+All declarations for literals, aggregates, operators forming an expressions.
+"""
 from typing import Tuple, List, Iterable
 
 from pyTooling.Decorators import export
 
-from pyVHDLModel import ModelEntity, ExpressionUnion, Direction, SubtypeOrSymbol, Symbol
+from pyVHDLModel import ModelEntity, ExpressionUnion, Direction, SubtypeOrSymbol
+from pyVHDLModel.Symbol import NewSymbol
 
 
 @export
@@ -565,16 +602,16 @@ class Allocation(BaseExpression):
 
 @export
 class SubtypeAllocation(Allocation):
-	_subtype: Symbol
+	_subtype: NewSymbol
 
-	def __init__(self, subtype: Symbol):
+	def __init__(self, subtype: NewSymbol):
 		super().__init__()
 
 		self._subtype = subtype
 		subtype._parent = self
 
 	@property
-	def Subtype(self) -> Symbol:
+	def Subtype(self) -> NewSymbol:
 		return self._subtype
 
 	def __str__(self) -> str:
@@ -665,16 +702,16 @@ class RangedAggregateElement(AggregateElement):
 
 @export
 class NamedAggregateElement(AggregateElement):
-	_name: Symbol
+	_name: NewSymbol
 
-	def __init__(self, name: Symbol, expression: ExpressionUnion):
+	def __init__(self, name: NewSymbol, expression: ExpressionUnion):
 		super().__init__(expression)
 
 		self._name = name
 		name._parent = self
 
 	@property
-	def Name(self) -> Symbol:
+	def Name(self) -> NewSymbol:
 		return self._name
 
 	def __str__(self) -> str:
