+++ pyVHDLModel/Root.py
@@ -1,19 +1,54 @@
-from pathlib import Path
-from typing import Dict, List, Generator, Union
+# ==================================================================================================================== #
+#             __     ___   _ ____  _     __  __           _      _                                                     #
+#   _ __  _   \ \   / / | | |  _ \| |   |  \/  | ___   __| | ___| |                                                    #
+#  | '_ \| | | \ \ / /| |_| | | | | |   | |\/| |/ _ \ / _` |/ _ \ |                                                    #
+#  | |_) | |_| |\ V / |  _  | |_| | |___| |  | | (_) | (_| |  __/ |                                                    #
+#  | .__/ \__, | \_/  |_| |_|____/|_____|_|  |_|\___/ \__,_|\___|_|                                                    #
+#  |_|    |___/                                                                                                        #
+# ==================================================================================================================== #
+# Authors:                                                                                                             #
+#   Patrick Lehmann                                                                                                    #
+#                                                                                                                      #
+# License:                                                                                                             #
+# ==================================================================================================================== #
+# Copyright 2017-2023 Patrick Lehmann - Boetzingen, Germany                                                            #
+# Copyright 2016-2017 Patrick Lehmann - Dresden, Germany                                                               #
+#                                                                                                                      #
+# Licensed under the Apache License, Version 2.0 (the "License");                                                      #
+# you may not use this file except in compliance with the License.                                                     #
+# You may obtain a copy of the License at                                                                              #
+#                                                                                                                      #
+#   http://www.apache.org/licenses/LICENSE-2.0                                                                         #
+#                                                                                                                      #
+# Unless required by applicable law or agreed to in writing, software                                                  #
+# distributed under the License is distributed on an "AS IS" BASIS,                                                    #
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.                                             #
+# See the License for the specific language governing permissions and                                                  #
+# limitations under the License.                                                                                       #
+#                                                                                                                      #
+# SPDX-License-Identifier: Apache-2.0                                                                                  #
+# ==================================================================================================================== #
+#
+"""
+This module contains parts of an abstract document language model for VHDL.
 
-from pyTooling.Decorators import export
-from pyTooling.Graph import Graph, Vertex
 
-from pyVHDLModel import ModelEntity, DesignUnit, DesignUnitKind, DependencyGraphVertexKind, DependencyGraphEdgeKind, PrimaryUnit, NamedEntityMixin, \
-	DocumentedEntityMixin
-from pyVHDLModel.Concurrent import EntityInstantiation, ComponentInstantiation, ConfigurationInstantiation
-from pyVHDLModel.DesignUnit import Architecture, PackageBody, Context, Entity, Configuration, Package
-from pyVHDLModel.Exception import LibraryExistsInDesignError, LibraryRegisteredToForeignDesignError, LibraryNotRegisteredError, EntityExistsInLibraryError, \
-	ArchitectureExistsInLibraryError, PackageExistsInLibraryError, PackageBodyExistsError, ConfigurationExistsInLibraryError, ContextExistsInLibraryError, \
-	ReferencedLibraryNotExistingError
-from pyVHDLModel.PSLModel import VerificationUnit, VerificationProperty, VerificationMode
-from pyVHDLModel.Symbol import AllPackageMembersReferenceSymbol, PackageMembersReferenceSymbol
-from pyVHDLModel.SyntaxModel import PackageInstantiation
+"""
+from pathlib                   import Path
+from typing                    import Dict, List, Generator, Union
+
+from pyTooling.Decorators      import export
+from pyTooling.Graph           import Graph, Vertex
+
+from pyVHDLModel               import ModelEntity, DesignUnitKind, DependencyGraphVertexKind, DependencyGraphEdgeKind, NamedEntityMixin, DocumentedEntityMixin
+from pyVHDLModel.Symbol        import AllPackageMembersReferenceSymbol, PackageMembersReferenceSymbol
+from pyVHDLModel.Concurrent    import EntityInstantiation, ComponentInstantiation, ConfigurationInstantiation
+from pyVHDLModel.DesignUnit    import Architecture, PackageBody, Context, Entity, Configuration, Package, DesignUnit, PrimaryUnit
+from pyVHDLModel.Exception     import LibraryExistsInDesignError, LibraryRegisteredToForeignDesignError, LibraryNotRegisteredError, EntityExistsInLibraryError
+from pyVHDLModel.Exception     import ArchitectureExistsInLibraryError, PackageExistsInLibraryError, PackageBodyExistsError, ConfigurationExistsInLibraryError
+from pyVHDLModel.Exception     import ContextExistsInLibraryError, ReferencedLibraryNotExistingError
+from pyVHDLModel.PSLModel      import VerificationUnit, VerificationProperty, VerificationMode
+from pyVHDLModel.Instantiation import PackageInstantiation
 
 
 @export
